[Feat] PlayerHealth Add
-PlayerHealth 스크립트 작성
플레이어의 피격을 구현하였으며, hp 회복, 최대hp 설정 및 플레이어 Die처리 완료

-추가 개발 사항: 플레이어 피격이 일정시간 무적을 구현하였습니다 이후 기획에서 필요없다고 하면 제거예정

+++ Assets/PMS/Scripts/Interface/IDamagedable.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    public void Damaged(float aktDamage);
+    public void Damaged(int aktDamage);
 }
