+++ Assets/PMS/Scripts/Weapon/Gun/GunWeaponBase.cs
@@ -9,7 +9,7 @@ public abstract class GunWeaponBase : MonoBehaviour
 
     //총 본체 관련 변수
     [SerializeField] protected string _gunName;     // 총의 이름
-    [SerializeField] protected float _damage;       // 총의 데미지
+    [SerializeField] protected int _damage;       // 총의 데미지
     [SerializeField] protected float _range;        // 총의 유효 사거리
     [SerializeField] protected float _accuracy;     // 총의 정확도
 
@@ -29,7 +29,7 @@ public abstract class GunWeaponBase : MonoBehaviour
     [SerializeField]protected AudioClip fire_Sound;        // 총 발사 소리 오디오 클립
     protected ObjectPool _gunBulletObjectPool;
 
-    public float Damage { get { return _damage; } private set { } }
+    public int Damage { get { return _damage; } private set { } }
     public virtual void Init()
     {
         //무기마다 Init() 불릿 풀 사이즈가 다를 것이고,불릿 프리팹이 다르다.
